refactor: add commands service and events file
