+++ src/commands/commands.constants.ts
@@ -1,0 +1,3 @@
+export enum Commands {
+  SUMMARY = '/summary',
+}
