+++ src/main.ts
@@ -6,7 +6,7 @@ import helmet, { contentSecurityPolicy } from 'helmet';
 import morgan from 'morgan';
 import { appRouter } from './app.router';
 import { dataSource } from './database/data-source';
-import { initMatrixEventHandlers } from './matrix/matrix.service';
+import { initMatrixEventHandlers } from './matrix/matrix.events';
 import { getOllamaInitMessage } from './ollama/ollama.service';
 
 dotenv.config();
